docs/requirements.txt
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
django-extensions
pydot
-r ../requirements.txt